Add checks to Matrix class operations
operator+: both matrices must have the same shape.

operator-: both matrices must have the same shape.

operator*: number of coloumns of first matrix must equal number of rows of the second matrix.

power: matrix must be square.

determinant: matrix must be square.

inverse: matrix must be square.

operator/: matrix must be square

+++ Matrix.h
@@ -45,7 +45,7 @@ public:
 
     Matrix transpose() const;
 
-    Matrix power(int n) const;
+    Matrix power(double n) const;
 
     Complex determinant() const;
